# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Swasthera Settlement Module API
servers:
  - url: /api
...
paths:
...
  /onboardings:
    get:
      operationId: listOnboardings
      tags: [onboarding]
      summary: List all onboarding submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, ACTIVE]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Onboarding list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingSummary"
    post:
      operationId: createOnboarding
      tags: [onboarding]
      summary: Create new onboarding draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnboardingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
  /onboardings/{id}:
    get:
      operationId: getOnboarding
      tags: [onboarding]
      summary: Get onboarding by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Onboarding detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
    put:
      operationId: updateOnboarding
      tags: [onboarding]
      summary: Update onboarding draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOnboardingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
  /onboardings/{id}/submit:
    post:
      operationId: submitOnboarding
      tags: [onboarding]
      summary: Maker submits onboarding for Checker review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
  /onboardings/{id}/approve:
    post:
      operationId: approveOnboarding
      tags: [onboarding]
      summary: Checker approves onboarding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckerDecisionBody"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
  /onboardings/{id}/reject:
    post:
      operationId: rejectOnboarding
      tags: [onboarding]
      summary: Checker rejects onboarding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckerDecisionBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Onboarding"
...
  /compliance/tcs-tds:
    get:
      operationId: getTcsTdsSummary
      tags: [compliance]
      summary: TCS/TDS register summary
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: TCS/TDS summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TcsTdsSummary"
  /compliance/tcs-records:
    get:
      operationId: listTcsRecords
      tags: [compliance]
      summary: TCS records by warehouse state GSTIN
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: TCS records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TcsRecord"
  /compliance/tds-records:
    get:
      operationId: listTdsRecords
      tags: [compliance]
      summary: TDS records by company TAN
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: TDS records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TdsRecord"
...
components:
  schemas:
...
    OnboardingSummary:
      type: object
      properties:
        id:
          type: integer
        ref:
          type: string
        companyName:
          type: string
        brandName:
          type: string
        companyType:
          type: string
        submittedBy:
          type: string
        submittedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, ACTIVE]
        kybStatus:
          type: string
          enum: [NOT_STARTED, PENDING, PASSED, FAILED]
        docsUploaded:
          type: integer
        docsRequired:
          type: integer
        urgency:
          type: string
          enum: [Normal, Overdue, High]
      required: [id, ref, companyName, brandName, companyType, status, kybStatus, docsUploaded, docsRequired, urgency]
    Onboarding:
      type: object
      properties:
        id:
          type: integer
        ref:
          type: string
        status:
          type: string
          enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, ACTIVE]
        kybStatus:
          type: string
          enum: [NOT_STARTED, PENDING, PASSED, FAILED]
        companyName:
          type: string
        companyType:
          type: string
        pan:
          type: string
        cin:
          type: string
        masterGstin:
          type: string
        tan:
          type: string
        bankAccount:
          type: string
        bankIfsc:
          type: string
        bankName:
          type: string
        spocName:
          type: string
        spocEmail:
          type: string
        spocMobile:
          type: string
        brandName:
          type: string
        brandCategory:
          type: string
        brandType:
          type: string
        warehouseName:
          type: string
        warehouseState:
          type: string
        warehouseGstin:
          type: string
        warehouseAddress:
          type: string
        commissionRate:
          type: number
        commissionType:
          type: string
        returnWindowDays:
          type: integer
        tcsRate:
          type: number
        tdsRate:
          type: number
        docsUploaded:
          type: integer
        docsRequired:
          type: integer
        submittedBy:
          type: string
        submittedAt:
          type: string
          format: date-time
        checkerName:
          type: string
        checkerNotes:
          type: string
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        fyndCompanyCode:
          type: string
        fyndBrandId:
          type: string
        fyndLocationId:
          type: string
      required: [id, ref, status, kybStatus, companyName, companyType, pan, masterGstin, bankAccount, bankIfsc, bankName, brandName, brandCategory, brandType, warehouseName, warehouseState, warehouseGstin, warehouseAddress, commissionRate, commissionType, returnWindowDays, tcsRate, tdsRate, docsUploaded, docsRequired, createdAt, updatedAt]
    CreateOnboardingBody:
      type: object
      properties:
        companyName:
          type: string
        companyType:
          type: string
        pan:
          type: string
        cin:
          type: string
        masterGstin:
          type: string
        tan:
          type: string
        bankAccount:
          type: string
        bankIfsc:
          type: string
        bankName:
          type: string
        spocName:
          type: string
        spocEmail:
          type: string
        spocMobile:
          type: string
        brandName:
          type: string
        brandCategory:
          type: string
        brandType:
          type: string
        warehouseName:
          type: string
        warehouseState:
          type: string
        warehouseGstin:
          type: string
        warehouseAddress:
          type: string
        commissionRate:
          type: number
        commissionType:
          type: string
        returnWindowDays:
          type: integer
        tcsRate:
          type: number
        tdsRate:
          type: number
      required: [companyName, companyType, pan, masterGstin, bankAccount, bankIfsc, bankName, brandName, brandCategory, brandType, warehouseName, warehouseState, warehouseGstin, warehouseAddress, commissionRate, commissionType, returnWindowDays, tcsRate, tdsRate]
    UpdateOnboardingBody:
      type: object
      properties:
        companyName:
          type: string
        companyType:
          type: string
        pan:
          type: string
        cin:
          type: string
        masterGstin:
          type: string
        tan:
          type: string
        bankAccount:
          type: string
        bankIfsc:
          type: string
        bankName:
          type: string
        spocName:
          type: string
        spocEmail:
          type: string
        spocMobile:
          type: string
        brandName:
          type: string
        brandCategory:
          type: string
        brandType:
          type: string
        warehouseName:
          type: string
        warehouseState:
          type: string
        warehouseGstin:
          type: string
        warehouseAddress:
          type: string
        commissionRate:
          type: number
        commissionType:
          type: string
        returnWindowDays:
          type: integer
        tcsRate:
          type: number
        tdsRate:
          type: number
    CheckerDecisionBody:
      type: object
      properties:
        notes:
          type: string
        rejectionReason:
          type: string
      required: []
...
    TcsTdsSummary:
      type: object
      properties:
        month:
          type: string
        year:
          type: integer
        tcsAccrued:
          type: number
        tcsPaid:
          type: number
        tdsDeducted:
          type: number
        gstr8Status:
          type: string
          enum: [Pending, Filed, Overdue]
        gstr8DueDate:
          type: string
        tcsPaymentDue:
          type: string
        tdsDepositDue:
          type: string
      required: [month, year, tcsAccrued, tcsPaid, tdsDeducted, gstr8Status, gstr8DueDate, tcsPaymentDue, tdsDepositDue]
    TcsRecord:
      type: object
      properties:
        id:
          type: integer
        stateGstin:
          type: string
        stateCode:
          type: string
        stateName:
          type: string
        brandName:
          type: string
        taxableSupply:
          type: number
        tcsRate:
          type: number
        tcsAmount:
          type: number
        status:
          type: string
          enum: [Accrued, Paid, Filed]
        paymentDueDate:
          type: string
      required: [id, stateGstin, stateCode, stateName, brandName, taxableSupply, tcsRate, tcsAmount, status, paymentDueDate]
    TdsRecord:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        tan:
          type: string
        grossPayment:
          type: number
        tdsRate:
          type: number
        tdsAmount:
          type: number
        netPaid:
          type: number
        status:
          type: string
          enum: [Pending, Deposited, Filed]
      required: [id, companyName, tan, grossPayment, tdsRate, tdsAmount, netPaid, status]
...